Added new CRUD endpoints and global features (#29)
* Create service updated based on functional requirements

* Created exception filter (#11)

- Installed dependencies
- ExceptionFilter created
- Validations on BadRequest instance
- Created CustomValidationPipe
- Implemented at app level

* Updated schema and new migration (#12)

* Created exception filter

- Installed dependencies
- ExceptionFilter created
- Validations on BadRequest instance
- Created CustomValidationPipe
- Implemented at app level

* Updated prisma schema

Updated to follow bussiness rules and delete templates and educative programs on cascade

* Created new migration

* Implementation of field validation for creation and updating of partial templates.
Add ID and foreign key validations.

* Update the language of messages in dto files

* Restriction when updating partial template status

* Subjects module added (#18)

* Added subject module

* Created subject module

- created service
- created controller

* Created controller

- Controller created and fully configured
- Service updated to implement try-catch in each request
- interface implemented for subjects service

* Added JSDoc

- added jsdoc to service
- added jsdoc to controller

* updated CreateTemplateDto with restrictions and custom messages

* updated UpdateTemplateDto with restrictions and custom messages

* added the validateRevisedById method to verify foreign ids and their documentation with JSDoc

* added validateResponsibleId method to verify foreign ids and their documentation with JSDoc

* added the validateAreaId method to verify foreign ids and their documentation with JSDoc

* The message returned by the validateId method was updated to Spanish

* updated the method for creating templates with foreign key validations, exception handling, and JSDoc documentation

* updated method to list all templates with exception handling and JSDoc documentation

* updated the method for listing a template with foreign key validations, exception handling, and JSDoc documentation

* The currentPeriod method was added to generate the current quarter and the documentation with JSDoc

* Updated the method for updating a template with foreign key validations, exception handling, and JSDoc documentation

* updated the method for deleting a template with foreign key validations, exception handling, and JSDoc documentation

* The validation messages for the updateDto file were modified

* Changed updateDto and createDto file validation messages

* Modified the findOne method to return the templates when they are found

* Created the complete CRUD resource for areas

* The area folder and the createDto and updateDto files were created within models

* Fixed simple syntax errors with Eslint

* Added validation to createDto with custom constraints and messages

* Added validation to updateDto with custom constraints and messages

* Updated the method to create area and the success or error messages

* Messages emitted by the create method have been updated

* Updated findOne method, added validations and error handling

* Updated findOne method name to findOneById

* The findOneByName method was created with exception and error handling

* Updated the create method with validations and exception handling

* Educational programs module added (#19)

* Validation when creating a new educational program

* Validation aggregation of information from an educational program.

* Changes made

* comments in spanish

* Changes made to the controller, to receive confirmation

* Removing unnecessary comments

* Changes made, I am waiting for responses

---------


* Updated the remove method with validations and exception handling

* The validateIfExistsAreaId method was created to validate if the area id exists, throwing a message if it is not found

* Restrict total hours by position

* Function to filter records by status

* The structured return with message, error and data in the response was implemented in the CRUD

* Message changed to Spanish

* Fixed unique constraint error on subject create (#22)

* Added subject module

* Created subject module

- created service
- created controller

* Created controller

- Controller created and fully configured
- Service updated to implement try-catch in each request
- interface implemented for subjects service

* Added JSDoc

- added jsdoc to service
- added jsdoc to controller

* Updated migration

- Created migration to be up to date with the database
- Updated validations

* updated CRUD operations and validation

- Updated validations
- Updated validaiotn in custom validation pipe

* Finally!! CRUD finished

- Added correct validations
- Fixed error on create "unique constraint violation"

* Updated DTO for update

- Added validations for update DTO

* Updating DTOs, controller and service files

* Added providers to PrismaService in module file

* The controller file was updated with the corresponding methods

* added name validation and id validation

* Updated the area query method to a single one in the controller file

* Updated the method to query the name of an area in addition to sorting them in ascending order

* Added documentation to the CRUD fot the Area in the service file

* Added documentation to CRUD for Areas in controller file

* Added id validations in service and controller files

* Elimination of the spread operator

* Updated the validate Foreign Keys method

* Hot Fix: inconsistency of changes since last merge conflicts resolution (#25)

* Added subject module

* Created subject module

- created service
- created controller

* Created controller

- Controller created and fully configured
- Service updated to implement try-catch in each request
- interface implemented for subjects service

* Added JSDoc

- added jsdoc to service
- added jsdoc to controller

* Updated migration

- Created migration to be up to date with the database
- Updated validations

* updated CRUD operations and validation

- Updated validations
- Updated validaiotn in custom validation pipe

* Finally!! CRUD finished

- Added correct validations
- Fixed error on create "unique constraint violation"

* Updated DTO for update

- Added validations for update DTO

* Hot fixes

- Updated validate method in validateforeignKeys
- Updated find method in controller, now uses query params and parse intPipe
- Fixed wrong name of validate() in service
-

* Added missing route in controller

- Added delete handler, now exposes the logic for http delete requests

* Revert "Implementation of restrictions, validations and custom messages"

* Revert "Revert "Implementation of restrictions, validations and custom messages"" (#28)

---------

+++ src/educational-programs/educational-programs.service.ts
@@ -1,14 +1,16 @@
-import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { CreateEducationalProgramDto } from '../models/EducationalPrograms/create-educational-program.dto';
 import { UpdateEducationalProgramDto } from '../models/EducationalPrograms/update-educational-program.dto';
 import { PrismaService } from 'src/prisma.service';
 import { EducationalPrograms, Prisma } from '@prisma/client';
 import { promises } from 'dns';
 
-
 @Injectable()
 export class EducationalProgramsService {
-
   constructor(private prisma: PrismaService) {}
 
 /**
@@ -16,45 +18,57 @@ export class EducationalProgramsService {
  * @param data Program data to register
  * @returns Registered Educational Program
  */
-  async createProg(data: Prisma.EducationalProgramsCreateInput): Promise<EducationalPrograms> {
-    const existing = await this.prisma.educationalPrograms.findFirst({
-      where:{
-        description: data.description,
-      },
-    });
-    if(existing){
-      throw new BadRequestException(`This educational program already exists`);
-    }
-   return this.prisma.educationalPrograms.create({
-    data,
-   })
+async createProgram(educationalProgram: CreateEducationalProgramDto): Promise <{ message: string | null; error: { message: string } | null; data: EducationalPrograms | null }> {
+  try {
+  await this.validateAreaId(educationalProgram);
+  const newPartialTemplate = await this.prisma.educationalPrograms.create({
+    data: {
+      ...educationalProgram,
+    },
+  });
+
+  return { message: 'Registrado con éxito', error: null, data: newPartialTemplate};
+} catch (error) {
+  return { message: 'No se pudo registrar el programa educativo', error: error.message , data: null };
   }
+ }
 
   /**
    * Method to consult all programs
    * @returns Returns all registered educational programs
    */
-  async findAll(): Promise<EducationalPrograms[]> {
-    return await this.prisma.educationalPrograms.findMany();
+  async findAllPrograms(): Promise<any> {
+    try {
+      const programs = await this.prisma.educationalPrograms.findMany();
+      
+      if (programs.length === 0) {
+        return { message: 'Sin programas educativos', error: null, data: null };
+      }
+      
+     
+      return { message: 'Programas educativos encontrados', error: null, data: programs };
+    } catch (error) {
+      return { message: 'Error al cargar programas educativos', error: error.message, data: null };
+    } 
   }
 
-  /**
-   * Method to query a programs from the id
-   * @param id id of the program to consult
-   * @returns Returns the program with the corresponding id
-   */
 
-  async byId(id: number): Promise<EducationalPrograms> {
-    const edu_project= await this.prisma.educationalPrograms.findUnique({
-      where:{
-        id,
-      },
+   /**
+   * Method to find an educational program by its id.
+   * @param id ID of the program to find
+   * @returns Educational program
+   * @throws NotFoundException if program is not found
+   */
+   async findProgramById(id: number): Promise<EducationalPrograms> {
+    const educationalProgram = await this.prisma.educationalPrograms.findUnique({
+      where: { id },
     });
 
-    if(!edu_project){
-      throw new NotFoundException (`Educational program not found`)
+    if (!educationalProgram) {
+      throw new NotFoundException(`Programa educativo con ID ${id} no encontrado`);
     }
-    return edu_project;
+
+    return educationalProgram;
   }
 
   /**
@@ -64,12 +78,26 @@ export class EducationalProgramsService {
    * @returns Returns the updated program
    */
 
-  async update(id: number, updateEducationalProgramDto: UpdateEducationalProgramDto,): Promise<EducationalPrograms> {
-    await this.byId(id);
-    return await this.prisma.educationalPrograms.update({data: {...updateEducationalProgramDto}as any, where: { id},
-    });
+  async updateProgram(
+    id: number,
+    updateEducationalProgramDto: UpdateEducationalProgramDto
+  ): Promise<{ message: string | null; error: { message: string } | null; data: EducationalPrograms | null }> {
+    try {
+      await this.validateAreaId(updateEducationalProgramDto);
   
+      await this.findProgramById(id);
+  
+      const updatedProgram = await this.prisma.educationalPrograms.update({
+        data: { ...updateEducationalProgramDto },
+        where: { id },
+      });
+  
+      return { message: 'Actualización exitosa', error: null, data: updatedProgram };
+    } catch (error) {
+      return { message: 'Error al actualizar', error: { message: error.message }, data: null };
+    }
   }
+  
   
 /**
  * *Method to delete a program
@@ -77,13 +105,47 @@ export class EducationalProgramsService {
  * @returns Return a message after deleting a program
  */
 
- async remove(id: number): Promise<{ message: string }> {
-  await this.byId(id);
+async removeProgram(id: number, confirmed: boolean): Promise<{ message: string }> {
+ 
+  if (!confirmed) {
+    return { message: 'Operación no confirmada por el usuario' };
+  }
+
+  
+  await this.findProgramById(id);
+
   await this.prisma.educationalPrograms.delete({
-    where: {
-      id
-    }
+    where: { id }
   });
-  return { message: 'Deleted successfully' };
-  }
+
+  return { message: 'Eliminada Correctamente' };
+}
+
+
+/**
+   * Validates if foreign keys (areaId, responsibleId, revisedById)
+   * @param {CreateEducationalProgramDto | UpdateEducationalProgramDto} dto - Data to validate
+   * @throws {NotFoundException} - If any foreign keys is not found
+   */
+private async validateAreaId(
+  dto: CreateEducationalProgramDto | UpdateEducationalProgramDto,
+): Promise<void> {
+  const { areaId } = dto;
+  const validations = [];
+
+  if (areaId !== undefined) {
+    if (typeof areaId !== 'number') {
+      throw new Error(`Tipo de areaId inválido: se esperaba un número, se recibió ${typeof areaId}`);
+    }
+    validations.push(
+      this.prisma.area.count({ where: { id: areaId } }).then((count) => {
+        if (count === 0) {
+          throw new NotFoundException(`Área con ID ${areaId} no existe`);
+        }
+      }),
+    );
+  }
+
+  await Promise.all(validations);
+}
 }
